refactor(estructure): Validators, interfaces and ReadModels factorization
