+++ src/Bio.Backend.Core/Bio.Application/DTOs/UserUpdateDTO.cs
@@ -20,6 +20,7 @@ public class UserUpdateDTO
     /// </summary>
     [Required(ErrorMessage = "Email is required.")]
     [EmailAddress(ErrorMessage = "Email format is invalid.")]
+    [StringLength(100, ErrorMessage = "Email cannot exceed 100 characters.")]
     public string Email { get; set; } = string.Empty;
 
     /// <summary>
