+++ src/Bio.Backend.Core/Bio.Application/Services/UserService.cs
@@ -2,7 +2,6 @@ using Bio.Application.DTOs;
 using Bio.Application.Interfaces;
 using Bio.Domain.Entities;
 using Bio.Domain.Interfaces;
-using FluentValidation;
 
 namespace Bio.Application.Services;
 
@@ -14,22 +13,16 @@ public class UserService : IUserService
 {
     private readonly IUserRepository _userRepository;
     private readonly IPasswordHasher _passwordHasher;
-    private readonly IValidator<UserCreateDTO> _validator;
-    private readonly IValidator<UserUpdateDTO> _updateValidator;
 
     /// <summary>
     /// Initializes a new instance of <see cref="UserService"/>.
     /// </summary>
     /// <param name="userRepository">User repository.</param>
     /// <param name="passwordHasher">Password hashing service for security.</param>
-    /// <param name="validator">Validator for user creation data.</param>
-    /// <param name="updateValidator">Validator for user update data.</param>
-    public UserService(IUserRepository userRepository, IPasswordHasher passwordHasher, IValidator<UserCreateDTO> validator, IValidator<UserUpdateDTO> updateValidator)
+    public UserService(IUserRepository userRepository, IPasswordHasher passwordHasher)
     {
         _userRepository = userRepository;
         _passwordHasher = passwordHasher;
-        _validator = validator;
-        _updateValidator = updateValidator;
     }
 
     /// <summary>
@@ -39,13 +32,6 @@ public class UserService : IUserService
     /// <returns>DTO with the data of the created user (Id, Name, Email, etc.).</returns>
     public async Task<UserResponseDTO> CreateUserAsync(UserCreateDTO userCreateDTO)
     {
-        // 0. Validation
-        var validationResult = await _validator.ValidateAsync(userCreateDTO);
-        if (!validationResult.IsValid)
-        {
-            throw new ValidationException(validationResult.Errors);
-        }
-
         // 0. Check if user already exists
         var existingUser = await _userRepository.GetByEmailAsync(userCreateDTO.Email);
         if (existingUser != null)
@@ -143,26 +129,21 @@ public class UserService : IUserService
     /// </summary>
     public async Task<UserResponseDTO?> UpdateUserAsync(Guid id, UserUpdateDTO dto)
     {
-        // 1. Validate DTO format
-        var validationResult = await _updateValidator.ValidateAsync(dto);
-        if (!validationResult.IsValid)
-            throw new ValidationException(validationResult.Errors);
-
-        // 2. Retrieve the user to update
+        // 1. Retrieve the user to update
         var user = await _userRepository.GetByIdAsync(id);
         if (user == null) return null;
 
-        // 3. Check uniqueness of email (excluding self)
+        // 2. Check uniqueness of email (excluding self)
         var emailConflict = await _userRepository.GetByEmailExcludingIdAsync(dto.Email, id);
         if (emailConflict != null)
             throw new InvalidOperationException($"User with email {dto.Email} already exists.");
 
-        // 4. Check uniqueness of phone (excluding self)
+        // 3. Check uniqueness of phone (excluding self)
         var phoneConflict = await _userRepository.GetByPhoneNumberExcludingIdAsync(dto.PhoneNumber, id);
         if (phoneConflict != null)
             throw new InvalidOperationException($"User with phone number {dto.PhoneNumber} already exists.");
 
-        // 5. Apply changes
+        // 4. Apply changes
         user.FullName = dto.FullName;
         user.Email = dto.Email;
         user.PhoneNumber = dto.PhoneNumber;
